# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `hr_manager`, `project_manager`, `team_leader`, `team_member`, `compliance_officer`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get a task by ID (includes subtasks and active escalations)",
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Actor user id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `team_member`, `team_leader`, `project_manager`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Update a task",
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Actor user id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `project_manager`, `team_leader`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Delete a task",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/subtasks": {
      "get": {
        "operationId": "SubtasksController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `project_manager`, `team_leader`, `team_member`, `compliance_officer`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all subtasks",
        "tags": [
          "Subtasks"
        ]
      },
      "post": {
        "operationId": "SubtasksController_create",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `team_leader`, `project_manager`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubtaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a subtask",
        "tags": [
          "Subtasks"
        ]
      }
    },
    "/subtasks/by-task/{taskId}": {
      "get": {
        "operationId": "SubtasksController_findByTask",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `project_manager`, `team_leader`, `team_member`, `compliance_officer`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get subtasks for a task",
        "tags": [
          "Subtasks"
        ]
      }
    },
    "/subtasks/{id}": {
      "get": {
        "operationId": "SubtasksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `project_manager`, `team_leader`, `team_member`, `compliance_officer`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a subtask by ID",
        "tags": [
          "Subtasks"
        ]
      },
      "patch": {
        "operationId": "SubtasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `team_member`, `team_leader`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubtaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a subtask",
        "tags": [
          "Subtasks"
        ]
      },
      "delete": {
        "operationId": "SubtasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `team_leader`, `project_manager`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a subtask",
        "tags": [
          "Subtasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "OfficeSync API",
    "description": "Backend API for the OfficeSync HR and PM Dashboard",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Role-Based-Access": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Enter your role to unlock the endpoints (e.g., platform_admin, superuser, hr_manager, team_leader, team_member)"
      }
    },
    "schemas": {
...
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "comp-uuid-123"
          },
          "title": {
            "type": "string",
            "example": "Reconcile vendor invoices"
          },
          "description": {
            "type": "string",
            "example": "Cross-check all vendor invoices against PO records."
          },
          "project_id": {
            "type": "string",
            "example": "proj-uuid-1"
          },
          "created_by": {
            "type": "string",
            "example": "user-uuid-4"
          },
          "assigned_to": {
            "type": "string",
            "example": "user-uuid-5"
          },
          "priority": {
            "type": "object",
            "example": "High"
          },
          "status": {
            "type": "object",
            "example": "Active"
          },
          "estimated_hours": {
            "type": "number",
            "example": 8
          },
          "actual_hours": {
            "type": "number",
            "example": 4.5
          },
          "due_date": {
            "type": "string",
            "example": "2024-11-20"
          },
          "processInstanceStepId": {
            "type": "string",
            "example": "step-uuid-10"
          }
        }
      },
      "CreateSubtaskDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "comp-uuid-123"
          },
          "task_id": {
            "type": "string",
            "example": "task-uuid-101"
          },
          "title": {
            "type": "string",
            "example": "Export invoice data from ERP"
          },
          "description": {
            "type": "string",
            "example": "Pull all Q4 invoices from SAP into a CSV."
          },
          "created_by": {
            "type": "string",
            "example": "user-uuid-4",
            "description": "Defaults to actor when omitted"
          },
          "assigned_to": {
            "type": "string",
            "example": "user-uuid-5"
          },
          "status": {
            "type": "object",
            "example": "Draft"
          },
          "estimated_hours": {
            "type": "number",
            "example": 2.5
          },
          "due_date": {
            "type": "string",
            "example": "2024-11-18"
          }
        },
        "required": [
          "companyId",
          "task_id",
          "title"
        ]
      },
      "UpdateSubtaskDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "comp-uuid-123"
          },
          "task_id": {
            "type": "string",
            "example": "task-uuid-101"
          },
          "title": {
            "type": "string",
            "example": "Export invoice data from ERP"
          },
          "description": {
            "type": "string",
            "example": "Pull all Q4 invoices from SAP into a CSV."
          },
          "created_by": {
            "type": "string",
            "example": "user-uuid-4",
            "description": "Defaults to actor when omitted"
          },
          "assigned_to": {
            "type": "string",
            "example": "user-uuid-5"
          },
          "status": {
            "type": "object",
            "example": "Draft"
          },
          "estimated_hours": {
            "type": "number",
            "example": 2.5
          },
          "due_date": {
            "type": "string",
            "example": "2024-11-18"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Role-Based-Access": []
    }
  ]
}